+++ modules/atomic-widgets/prop-types/concerns/has-atomic-base.php
@@ -5,6 +5,10 @@ namespace Elementor\Modules\AtomicWidgets\PropTypes\Concerns;
 use Elementor\Modules\AtomicWidgets\Base\Atomic_Control_Base;
 use Elementor\Modules\AtomicWidgets\Controls\Section;
 use Elementor\Modules\AtomicWidgets\PropsResolver\Props_Resolver;
+use Elementor\Modules\AtomicWidgets\PropTypes\Contracts\Prop_Type;
+use Elementor\Modules\AtomicWidgets\Styles\Style_Schema;
+use Elementor\Modules\AtomicWidgets\Validators\Props_Validator;
+use Elementor\Modules\AtomicWidgets\Validators\Style_Validator;
 use Elementor\Utils;
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -52,6 +56,42 @@ trait Has_Atomic_Base {
 		}
 
 		return $valid_controls;
+	}
+
+	private static function validate_schema( array $schema ) {
+		$widget_name = static::class;
+
+		foreach ( $schema as $key => $prop ) {
+			if ( ! ( $prop instanceof Prop_Type ) ) {
+				Utils::safe_throw( "Prop `$key` must be an instance of `Prop_Type` in `{$widget_name}`." );
+			}
+		}
+	}
+
+	private function sanitize_atomic_styles( array $styles ) {
+		foreach ( $styles as $style ) {
+			[$is_valid, $sanitized, $errors_bag] = Style_Validator::make( Style_Schema::get() )->validate( $style );
+
+			if ( ! $is_valid ) {
+				throw new \Exception( 'Styles validation failed. Invalid keys: ' . join( ', ', $errors_bag ) );
+			}
+
+			$styles[ $sanitized['id'] ] = $sanitized;
+		}
+
+		return $styles;
+	}
+
+	private function sanitize_atomic_settings( array $settings ): array {
+		$schema = static::get_props_schema();
+
+		[ , $validated, $errors ] = Props_Validator::make( $schema )->validate( $settings );
+
+		if ( ! empty( $errors ) ) {
+			throw new \Exception( 'Settings validation failed. Invalid keys: ' . join( ', ', $errors ) );
+		}
+
+		return $validated;
 	}
 
 	public function get_atomic_controls() {
